fix: make collab title fields optional, remove userid req
