+++ apps/api/drizzle/0010_make_collaborators_postition_optional.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "collaborators" ALTER COLUMN "position_title" DROP NOT NULL;
