+++ apps/ui/src/components/pages/application/modals/AddCollaboratorModal.tsx
@@ -134,7 +134,6 @@ const AddCollaboratorModal = memo(({ isOpen, setIsOpen }: ModalStateProps) => {
 									style={{ marginTop: '27px' }} // accounting for sublabel extra size from primaryEmail
 									control={control}
 									rule={rule}
-									required
 									disabled={!isEditMode}
 								/>
 							</Col>
